changeActivityStatus循环调用问题
